requirements.txt
# For parsing unstructured data (od-parse)
git+https://github.com/octondata/od-parse.git

# For NLP and generation (removed heavy deps to avoid Rust build on Py 3.13)

# For document retrieval (e.g., FAISS)
faiss-cpu==1.10.0

# For working with configuration files (YAML or JSON)
pyyaml==6.0.2

# For handling PDFs
PyMuPDF==1.24.11

# For working with images (if parsing images)
Pillow==10.4.0
pytesseract==0.3.13

# Other useful libraries
numpy==2.1.3
requests==2.32.3
